+++ .circleci/unittest/linux/scripts/environment.yml
@@ -26,4 +26,3 @@ dependencies:
     - mlflow
     - av
     - coverage
-    - ray
