+++ test/test_collector.py
@@ -31,7 +31,6 @@ from torchrl.collectors.collectors import (
     MultiSyncDataCollector,
     RandomPolicy,
 )
-from torchrl.collectors.distributed.ray import RayCollector
 from torchrl.collectors.utils import split_trajectories
 from torchrl.data import CompositeSpec, UnboundedContinuousTensorSpec
 from torchrl.envs import EnvCreator, ParallelEnv, SerialEnv, StepCounter
@@ -1230,68 +1229,6 @@ def test_initial_obs_consistency(env_class, seed=1):
 def weight_reset(m):
     if isinstance(m, nn.Conv2d) or isinstance(m, nn.Linear):
         m.reset_parameters()
-
-
-class TestRayDataCollector:
-    @pytest.mark.parametrize("frames_per_batch", [50, 100])
-    def test_ray_distributed_collector_basic(self, frames_per_batch):
-        env = ContinuousActionVecMockEnv()
-        policy = RandomPolicy(env.action_spec)
-        collector = RayCollector(
-            [env],
-            policy,
-            total_frames=1000,
-            frames_per_batch=frames_per_batch,
-        )
-        total = 0
-        for data in collector:
-            total += data.numel()
-            assert data.numel() == frames_per_batch
-        collector.shutdown()
-        assert total == 1000
-
-    @pytest.mark.parametrize("sync", [True, False])
-    def test_ray_distributed_collector_sync(self, sync):
-        frames_per_batch = 50
-        env = ContinuousActionVecMockEnv()
-        policy = RandomPolicy(env.action_spec)
-        collector = RayCollector(
-            [env],
-            policy,
-            total_frames=200,
-            frames_per_batch=frames_per_batch,
-            sync=sync,
-        )
-        total = 0
-        for data in collector:
-            total += data.numel()
-            assert data.numel() == frames_per_batch
-        collector.shutdown()
-        assert total == 200
-
-    @pytest.mark.parametrize("frames_per_batch", [50, 100])
-    @pytest.mark.parametrize(
-        "collector_class",
-        [SyncDataCollector, MultiaSyncDataCollector, MultiSyncDataCollector],
-    )
-    def test_ray_distributed_collector_collector_class(
-        self, frames_per_batch, collector_class
-    ):
-        env = ContinuousActionVecMockEnv()
-        policy = RandomPolicy(env.action_spec)
-        collector = RayCollector(
-            [env],
-            policy,
-            collector_class=collector_class,
-            total_frames=200,
-            frames_per_batch=frames_per_batch,
-        )
-        total = 0
-        for data in collector:
-            total += data.numel()
-            assert data.numel() == frames_per_batch
-        collector.shutdown()
-        assert total == 200
 
 
 @pytest.mark.skipif(_os_is_osx, reason="Queue.qsize does not work on osx.")
